requirements.txt
apache-airflow<2
requests